[refactor] Co-locate rotation snap behavior
Keep the field-specific interaction helper beside its sole production consumer instead of maintaining a standalone UI file.
